Use skin and unlock screen for UI tests
UI testing frameworks like Google's Espresso depend on how the UI
actually looks on the emulator. Starting the emulator with -no-skin
screws up the app's UI, and causes tests to fail. Also, the screen needs
to be unlocked for some kinds of tests to run, for example, those
that simulate keypresses[0] or wait for UI elements to appear on the
screen. Use a hack[1] to unlock the screen before starting tests.

[0] http://developer.android.com/tools/testing/activity_testing.html#UnlockDevice
[1] http://stackoverflow.com/a/3366443